refactor: 1차 프로젝트 구조 리펙토링
