+++ feature/feature-home/src/main/java/com/comit/feature_home/screen/schedule/ShowScheduleScreen.kt
@@ -62,6 +62,7 @@ import com.comit.feature_home.calendar.SimTongCalendar
 import com.comit.feature_home.getEndAt
 import com.comit.feature_home.getStartAt
 import com.comit.feature_home.mvi.FetchScheduleSideEffect
+import com.comit.feature_home.vm.ShowScheduleViewModel
 import com.comit.navigator.SimTongScreen
 import com.example.feature_home.R
 import kotlinx.coroutines.InternalCoroutinesApi
