+++ feature/feature-home/src/main/java/com/comit/feature_home/vm/WriteScheduleViewModel.kt
@@ -1,4 +1,4 @@
-package com.comit.feature_home.screen.schedule
+package com.comit.feature_home.vm
 
 import androidx.lifecycle.ViewModel
 import com.comit.domain.exception.BadRequestException
