+++ feature/feature-mypage/src/main/java/com/comit/feature_mypage/vm/InputCertificationViewModel.kt
@@ -1,4 +1,4 @@
-package com.comit.feature_mypage.screen.fix.email
+package com.comit.feature_mypage.vm
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
