+++ feature/feature-mypage/src/main/java/com/comit/feature_mypage/vm/MyPageViewModel.kt
@@ -1,4 +1,4 @@
-package com.comit.feature_mypage.screen
+package com.comit.feature_mypage.vm
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
